Core: Reset compromised passwords after 2FA failures
After a reasonable number of 2nd-factor attempts, it's safe to assume that the password been compromised and an attacker is trying to brute force the 2nd factor.

Many 2nd factors -- like TOTP and backup codes -- can be brute-forced in a matter of days, so some kind of rate-limiting is essential to provide a meaningful 2nd factor. Since we know that an attacker should never be able get this far in the first place, it's better to reset the password than to throttle attempts.